Moved `tracked_property` and `is_set` from `util` to `_helpers`
